Feature-homepage-localization: Adds translation for the case studies page
