+++ apps/web/messages/ar.json
@@ -121,6 +121,11 @@
     "view_case_study": "عرض دراسة الحالة",
     "view_all": "عرض جميع دراسات الحالة"
   },
+  "case_studies_page": {
+    "title": "دراسات الحالة الخاصة بنا",
+    "description": "استكشف محفظة مشاريعنا الكاملة الناجحة وشاهد كيف ساعدنا الشركات في تحويل حضورها الرقمي",
+    "sub_description": "من الشركات الناشئة المبتكرة إلى المؤسسات الراسخة، قدمنا حلولاً مخصصة تحفز النمو وتعزز تجربة المستخدم وتخلق تأثيرًا دائمًا"
+  },
   "services_section": {
     "what_we_do": "ماذا نفعل",
     "title": "خدمات تحقق النتائج",
